Move class 'Configuration' into new header 'configuration.hpp'
